+++ tests/integ/test_data_source_api.py
@@ -11,7 +11,6 @@ import datetime
 from unittest import mock
 from unittest.mock import patch, MagicMock, PropertyMock
 
-import oracledb
 import pytest
 from decimal import Decimal
 
@@ -20,13 +19,9 @@ from snowflake.snowpark._internal.data_source.datasource_partitioner import (
     DataSourcePartitioner,
 )
 from snowflake.snowpark._internal.data_source.datasource_reader import DataSourceReader
-from snowflake.snowpark._internal.data_source.drivers.oracledb_driver import (
-    output_type_handler,
-)
 from snowflake.snowpark._internal.data_source.drivers import (
     PyodbcDriver,
     SqliteDriver,
-    OracledbDriver,
 )
 from snowflake.snowpark._internal.data_source.utils import (
     _task_fetch_data_from_source_with_retry,
@@ -68,14 +63,12 @@ from tests.resources.test_data_source_dir.test_data_source_data import (
     sql_server_all_type_schema,
     SQLITE3_DB_CUSTOM_SCHEMA_STRING,
     SQLITE3_DB_CUSTOM_SCHEMA_STRUCT_TYPE,
-    oracledb_real_data,
     sql_server_udtf_ingestion_data,
     create_connection_oracledb,
     sql_server_create_connection_unicode_data,
     sql_server_create_connection_double_quoted_data,
-    oracledb_real_data_small,
 )
-from tests.utils import RUNNING_ON_JENKINS, Utils, IS_WINDOWS
+from tests.utils import Utils, IS_WINDOWS
 
 try:
     import pandas  # noqa: F401
@@ -112,29 +105,9 @@ def test_dbapi_with_temp_table(session, caplog):
         assert df.collect() == sql_server_all_type_data
 
 
-def test_dbapi_oracledb(session):
-    df = session.read.dbapi(
-        create_connection_oracledb,
-        table=ORACLEDB_TABLE_NAME,
-        max_workers=4,
-        query_timeout=5,
-    ).order_by("ID")
-    Utils.check_answer(df, oracledb_real_data)
-
-
 @pytest.mark.parametrize(
     "create_connection, table_name, expected_result",
     [
-        (
-            create_connection_oracledb,
-            ORACLEDB_TABLE_NAME,
-            oracledb_real_data,
-        ),
-        (
-            create_connection_oracledb,
-            ORACLEDB_TABLE_NAME_SMALL,
-            oracledb_real_data_small,
-        ),
         (sql_server_create_connection, SQL_SERVER_TABLE_NAME, sql_server_all_type_data),
         (
             sql_server_create_connection_small_data,
@@ -599,11 +572,6 @@ def test_type_conversion():
             sql_server_create_connection, DBMS_TYPE.SQL_SERVER_DB
         ).to_snow_type([("test_col", invalid_type, None, None, 0, 0, True)])
 
-    with pytest.raises(NotImplementedError, match="oracledb type not supported"):
-        OracledbDriver(create_connection_oracledb, DBMS_TYPE.ORACLE_DB).to_snow_type(
-            [invalid_type]
-        )
-
 
 def test_custom_schema_false(session):
     with pytest.raises(ValueError, match="Invalid schema string: timestamp_tz."):
@@ -785,45 +753,6 @@ def test_empty_table(session):
     assert df.collect() == []
 
 
-def test_oracledb_driver_coverage(caplog):
-    oracledb_driver = OracledbDriver(create_connection_oracledb, DBMS_TYPE.ORACLE_DB)
-    conn = oracledb_driver.prepare_connection(oracledb_driver.create_connection(), 0)
-    assert conn.outputtypehandler == output_type_handler
-
-    oracledb_driver.to_snow_type(
-        [OracleDBType("NUMBER_COL", oracledb.DB_TYPE_NUMBER, 40, 2, True)]
-    )
-    assert "Snowpark does not support column" in caplog.text
-
-
-@pytest.mark.skipif(
-    RUNNING_ON_JENKINS, reason="Cannot connect to oracledb from jenkins"
-)
-def test_udtf_ingestion_oracledb(session):
-    from tests.parameters import ORACLEDB_CONNECTION_PARAMETERS
-
-    def create_connection_oracledb():
-        import oracledb
-
-        host = ORACLEDB_CONNECTION_PARAMETERS["host"]
-        port = ORACLEDB_CONNECTION_PARAMETERS["port"]
-        service_name = ORACLEDB_CONNECTION_PARAMETERS["service_name"]
-        username = ORACLEDB_CONNECTION_PARAMETERS["username"]
-        password = ORACLEDB_CONNECTION_PARAMETERS["password"]
-        dsn = f"{host}:{port}/{service_name}"
-        connection = oracledb.connect(user=username, password=password, dsn=dsn)
-        return connection
-
-    df = session.read.dbapi(
-        create_connection_oracledb,
-        table="ALL_TYPE_TABLE",
-        udtf_configs={
-            "external_access_integration": ORACLEDB_TEST_EXTERNAL_ACCESS_INTEGRATION
-        },
-    ).order_by("ID")
-    Utils.check_answer(df, oracledb_real_data)
-
-
 def test_sql_server_udtf_ingestion(session):
     raw_schema = [
         ("Id", int, None, None, 10, 0, False),
@@ -925,16 +854,6 @@ def test_sql_server_udtf_ingestion(session):
         packages=["pyodbc"],
     )
     Utils.check_answer(df, sql_server_udtf_ingestion_data)
-
-
-def test_external_access_integration_not_set(session):
-    with pytest.raises(
-        ValueError,
-        match="external_access_integration cannot be None when udtf ingestion is used.",
-    ):
-        session.read.dbapi(
-            create_connection_oracledb, table=ORACLEDB_TABLE_NAME, udtf_configs={}
-        )
 
 
 def test_unknown_driver_with_custom_schema(session):
